weird error with converters __init malformed
